refactor(slots): add DownloadSlots interface
